move code to handler
